add step-4
